store: Move `EthereumCallCache` to `ChainStore`
Ultimately, we want the call cache to live next to the block cache for each
chain
